Updated Attack Hitboxes + New Attack Script
- Changed Player attack animations from a blend tree to a sub-state machine.
- Created PlayerAttacks script.
- Moved all attack-related variables and methods in the PlayerController script to the PlayerAttacks script except for attack damage.
- Changed attack animation events from separate methods to a single method with an enum switch.
