+++ Assets/Scenes/Test_PlayerAttackFunctionality.unity
@@ -21308,6 +21308,7 @@ GameObject:
   - component: {fileID: 1719131398}
   - component: {fileID: 1719131397}
   - component: {fileID: 1719131396}
+  - component: {fileID: 1719131401}
   - component: {fileID: 1719131394}
   m_Layer: 6
   m_Name: Player
@@ -21356,6 +21357,7 @@ MonoBehaviour:
   attackDamage: 1
   startingHealth: 100
   currentHealth: 0
+  isInShop: 0
 --- !u!114 &1719131397
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -21466,6 +21468,18 @@ Transform:
   - {fileID: 886736619}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1719131401
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1719131393}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c8f1cf54cd1bc8f41b1ad7fab834ecb5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1771263377
 GameObject:
   m_ObjectHideFlags: 0
