+++ Assets/Scripts/Player/PlayerAttacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8f1cf54cd1bc8f41b1ad7fab834ecb5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
