+++ Assets/Scripts/Player/PlayerController.cs
@@ -9,8 +9,6 @@ public class PlayerController : MonoBehaviour
     // Get reference variables
     private Rigidbody2D rb;
     private Animator animator;
-    private Camera mainCamera;
-    private Transform attackHitbox;
     public HealthBar healthBar;
 
     // Movement variables
@@ -18,44 +16,33 @@ public class PlayerController : MonoBehaviour
     private float idleSlow = 0.9f;
     public float startingMoveSpeed = 6f;
     [HideInInspector] public float moveSpeed;
-    
+
     // Attack variables
     public float startingAttackDamage = 10f;
     [HideInInspector] public float attackDamage;
-
-    private float attackCooldown = 0.3f;
-    private bool canAttack = true;
-
-    private Vector2 mousePos;
-    private Vector2 direction;
 
     // Health variables
     public int startingHealth = 100;
     public int currentHealth;
 
-    public bool isInShop { private get; set; } = false;
+    public bool isInShop = false;
 
     // Start is called before the first frame update
     void Start()
     {
         rb = GetComponent<Rigidbody2D>();
         animator = GetComponent<Animator>();
-        mainCamera = Camera.main;
-        attackHitbox = transform.GetChild(2);
 
         currentHealth = startingHealth;
         healthBar.SetHealth(currentHealth, startingHealth);
 
-        moveSpeed = startingMoveSpeed;
         attackDamage = startingAttackDamage;
+        moveSpeed = startingMoveSpeed;
     }
 
     // Update is called once per frame
     void Update()
     {
-        // Get mouse position
-        mousePos = mainCamera.ScreenToWorldPoint(Input.mousePosition);
-
         // Update animator
         bool isMoving = moveInput != Vector2.zero;
 
@@ -119,75 +106,5 @@ public class PlayerController : MonoBehaviour
         //}
 
         moveInput = inputValue.Get<Vector2>();
-    }
-
-    private void OnAttack()
-    {
-        if (!isInShop)
-        {
-            //if (canAttack)
-            //{
-            //    Weapon weapon = GetComponentInChildren<Weapon>();
-            //    weapon.animator.SetTrigger("Attack");
-
-            //    canAttack = false;
-            //    StartCoroutine(AttackCooldown());
-            //}
-
-            if (canAttack)
-            {
-                Vector2 direction = (mousePos - (Vector2)transform.position).normalized;
-                
-                Debug.Log("Horizontal value is: " + direction.x);
-                Debug.Log("Vertical value is: " + direction.y);
-                
-                animator.SetFloat("AttackHorizontal", direction.x);
-                animator.SetFloat("AttackVertical", direction.y);
-                animator.SetTrigger("Attack");
-
-                canAttack = false;
-                StartCoroutine(AttackCooldown());
-            }
-        }
-    }
-
-    public void AttackRightHitbox()
-    {
-        attackHitbox.localRotation = Quaternion.identity;
-        attackHitbox.localPosition = Vector2.zero;
-    }
-
-    public void AttackLeftHitbox()
-    {
-        Debug.Log("AttackLeftHitbox called.");
-        attackHitbox.localRotation = Quaternion.Euler(0, 180, 0);
-        attackHitbox.localPosition = Vector2.zero;
-    }
-
-    public void AttackUpHitbox()
-    {
-        attackHitbox.localRotation = Quaternion.Euler(0, 90, 0);
-        attackHitbox.localPosition = new Vector2(0.063f, 0.063f);
-    }
-
-    public void AttackDownHitbox()
-    {
-        attackHitbox.localRotation = Quaternion.Euler(0, 270, 0);
-        attackHitbox.localPosition = new Vector2(-0.189f, -0.189f);
-    }
-
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.gameObject.CompareTag("Enemy"))
-        {
-            Enemy enemy = collision.gameObject.GetComponent<Enemy>();
-            enemy.TakeDamage(attackDamage);
-        }
-    }
-
-    private IEnumerator AttackCooldown()
-    {
-        yield return new WaitForSeconds(attackCooldown);
-        canAttack = true;
     }
 }
